+++ sdk/communication/Azure.Communication.Chat/CHANGELOG.md
@@ -5,7 +5,7 @@
 - Added support for metadata in messages.
 - Added new overloads for sendmessage and updatemessage operations with options : SendChatMessageOptions and UpdateChatMessageOptions. 
 - Added support for user display name in typing notifications.
-- Added new overloads for sendtypingnotification operation with options : SendTypingNotificationOptions. 
+- Added new overloads for sendtypingnotification operation with options : TypingNotificationOptions. 
 
 ## 1.0.1 (2021-05-11)